backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
slowapi>=0.1.9
httpx>=0.27.0
psycopg2-binary>=2.9.9
